structure pyleri parser
