--- Dalaran/lexer/__init__.py
@@ -1,2 +1,0 @@
-from .lexer import lex
-from .tokens import token_exprs
