--- Dalaran/parser/utils.py
@@ -1,21 +1,0 @@
-class bidirectional_iterator(object):
-    def __init__(self, collection):
-        self.collection = collection
-        self.index = 0
-
-    def next(self):
-        try:
-            result = self.collection[self.index]
-            self.index += 1
-        except IndexError:
-            raise StopIteration
-        return result
-
-    def prev(self):
-        self.index -= 1
-        if self.index < 0:
-            raise StopIteration
-        return self.collection[self.index]
-
-    def __iter__(self):
-        return self
