+++ requirements.txt
@@ -1,0 +1,1 @@
+pyleri==1.3.0
